Split javascript interface methods into domain objects.
