+++ starboard/android/apk/app/src/main/java/dev/cobalt/util/SystemPropertiesHelper.java
@@ -16,36 +16,54 @@ package dev.cobalt.util;
 
 import static dev.cobalt.util.Log.TAG;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
 import java.lang.reflect.Method;
 
-/** Utility class for accessing system properties via reflection. */
 public class SystemPropertiesHelper {
-  private static Method getStringMethod;
+    private static Method getStringMethod;
 
-  static {
-    try {
-      getStringMethod =
-          ClassLoader.getSystemClassLoader()
-              .loadClass("android.os.SystemProperties")
-              .getMethod("get", String.class);
-      if (getStringMethod == null) {
-        Log.e(TAG, "Couldn't load system properties getString");
-      }
-    } catch (Exception exception) {
-      Log.e(TAG, "Exception looking up system properties methods: ", exception);
+    static {
+        try {
+            getStringMethod =
+                    ClassLoader.getSystemClassLoader()
+                            .loadClass("android.os.SystemProperties")
+                            .getMethod("get", String.class);
+            if (getStringMethod == null) {
+                Log.e(TAG, "Couldn't load system properties getString");
+            }
+        } catch (Exception exception) {
+            Log.e(TAG, "Exception looking up system properties methods: ", exception);
+        }
     }
-  }
 
-  private SystemPropertiesHelper() {}
+    private SystemPropertiesHelper() {}
 
-  public static String getString(String property) {
-    if (getStringMethod != null) {
-      try {
-        return (String) getStringMethod.invoke(null, new Object[] {property});
-      } catch (Exception exception) {
-        Log.e(TAG, "Exception getting system property: ", exception);
-      }
+    public static String getString(String property) {
+        if (getStringMethod != null) {
+            try {
+                return (String) getStringMethod.invoke(null, new Object[] {property});
+            } catch (Exception exception) {
+                Log.e(TAG, "Exception getting system property: ", exception);
+            }
+        }
+        return null;
     }
-    return null;
-  }
+
+    public static String getRestrictedSystemProperty(String propName, String defaultValue) {
+        try {
+            Process process = Runtime.getRuntime().exec("getprop " + propName);
+            BufferedReader bufferedReader =
+                    new BufferedReader(new InputStreamReader(process.getInputStream()));
+
+            return bufferedReader.readLine();
+        } catch (IOException e) {
+            return defaultValue;
+        }
+    }
+
+    public static String getSystemProperty(String propertyName, String defaultValue) {
+        return System.getProperty(propertyName, defaultValue);
+    }
 }
